\name{gatherEXPANDSoutput}
\alias{gatherEXPANDSoutput}
\title{Reading EXPANDS' output files}
\description{Reads EXPANDS output files from user-specified input directory.}
\usage{gatherEXPANDSoutput(outDirEXPANDS,regex="")}
\arguments{
  \item{outDirEXPANDS}{
    Absolute path to input directory in which EXPANDS results are stored, as generated by \code{\link{runExPANdS}}.
  }
  \item{regex}{
    Pattern that path has to match in order to be read.
  }
}

\value{Two-level nested list. The outer level contains one entry per each output of \code{\link{runExPANdS}}. Each entry constitutes an inner list with fields:
\cr
	\bold{snv} - The assignment of each SNV to a subpopulation. \cr
	\bold{cbs} - The copy number of each genomic segment in each subpopulation. \cr
	\bold{spstats} -  Matrix of predicted subpopulations. \cr
	\bold{tree} -  The inferred phylogenetic relationships between subpopulations as an object of class "phylo4" (library phylobase). \cr
	\bold{treeApe} - The inferred phylogenetic relationships between subpopulations as an object of class "phylo" (library ape).
}

\author{Noemi Andor}


\seealso{
  \code{\link{runExPANdS}}
}

